affichage map avec les brasseries de l'API
