+++ client/src/pages/Home.tsx
@@ -3,7 +3,7 @@ import "../components/home.css";
 import BeerContext from "../Contexts/BeerContext";
 
 import { useContext } from "react";
-import BreweriesList from "../components/BreweriesList";
+// import BreweriesList from "../components/BreweriesList";
 import Filter from "../components/Filter";
 import BoutonSoif from "../components/boutonSoif";
 import Maps from "../components/map";
@@ -24,9 +24,7 @@ function Home() {
       </div>
       <BoutonSoif />
       <Filter />
-      <div className="cardContainer">
-        <BreweriesList />
-      </div>
+      <div className="cardContainer">{/* <BreweriesList /> */}</div>
     </>
   );
 }
